refactor(agents): extract schedules into agents/schedules
