+++ packages/agents/src/tests/agents/migration.ts
@@ -73,77 +73,10 @@ export class TestMigrationAgent extends Agent {
 
   /**
    * Run the column-add + CHECK-constraint migration logic.
-   * This is the same code that runs in the Agent constructor.
+   * This is the same code the schedule component runs during lifecycle startup.
    */
   async runMigration(): Promise<void> {
-    // Step 1: Add missing columns (same as constructor)
-    const addColumnIfNotExists = (sql: string) => {
-      try {
-        this.ctx.storage.sql.exec(sql);
-      } catch (e) {
-        const message = e instanceof Error ? e.message : String(e);
-        if (!message.toLowerCase().includes("duplicate column")) {
-          throw e;
-        }
-      }
-    };
-
-    addColumnIfNotExists(
-      "ALTER TABLE cf_agents_schedules ADD COLUMN intervalSeconds INTEGER"
-    );
-    addColumnIfNotExists(
-      "ALTER TABLE cf_agents_schedules ADD COLUMN running INTEGER DEFAULT 0"
-    );
-    addColumnIfNotExists(
-      "ALTER TABLE cf_agents_schedules ADD COLUMN execution_started_at INTEGER"
-    );
-    addColumnIfNotExists(
-      "ALTER TABLE cf_agents_schedules ADD COLUMN retry_options TEXT"
-    );
-
-    // Step 2: Recreate table if CHECK constraint is missing 'interval'
-    const rows = this.ctx.storage.sql
-      .exec(
-        "SELECT sql FROM sqlite_master WHERE type='table' AND name='cf_agents_schedules'"
-      )
-      .toArray();
-    if (rows.length > 0) {
-      const ddl = String(rows[0].sql);
-      if (!ddl.includes("'interval'")) {
-        // Drop any leftover temp table from a previous partial migration
-        this.ctx.storage.sql.exec(
-          "DROP TABLE IF EXISTS cf_agents_schedules_new"
-        );
-        this.ctx.storage.sql.exec(`
-          CREATE TABLE cf_agents_schedules_new (
-            id TEXT PRIMARY KEY NOT NULL DEFAULT (randomblob(9)),
-            callback TEXT,
-            payload TEXT,
-            type TEXT NOT NULL CHECK(type IN ('scheduled', 'delayed', 'cron', 'interval')),
-            time INTEGER,
-            delayInSeconds INTEGER,
-            cron TEXT,
-            intervalSeconds INTEGER,
-            running INTEGER DEFAULT 0,
-            created_at INTEGER DEFAULT (unixepoch()),
-            execution_started_at INTEGER,
-            retry_options TEXT
-          )
-        `);
-        this.ctx.storage.sql.exec(`
-          INSERT INTO cf_agents_schedules_new
-            (id, callback, payload, type, time, delayInSeconds, cron,
-             intervalSeconds, running, created_at, execution_started_at, retry_options)
-          SELECT id, callback, payload, type, time, delayInSeconds, cron,
-                 intervalSeconds, running, created_at, execution_started_at, retry_options
-          FROM cf_agents_schedules
-        `);
-        this.ctx.storage.sql.exec("DROP TABLE cf_agents_schedules");
-        this.ctx.storage.sql.exec(
-          "ALTER TABLE cf_agents_schedules_new RENAME TO cf_agents_schedules"
-        );
-      }
-    }
+    this.schedules.ensureSchema();
   }
 
   /**
